# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for API e-mosque endpoints",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/profile": {
            "get": {
                "summary": "Get user profile",
                "description": "Retrieve the profile data of the authenticated user.",
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of user profile",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the profile retrieval",
                                            "type": "string",
                                            "example": "data user berhasil didapatkan"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User data not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the profile retrieval failure",
                                            "type": "string",
                                            "example": "data user tidak tersedia"
                                        },
                                        "data": {
                                            "description": "Additional error data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "failed"
                                        },
                                        "message": {
                                            "description": "Message indicating the status of the profile retrieval failure",
                                            "type": "string",
                                            "example": "Gagal menjalankan requesst"
                                        },
                                        "data": {
                                            "description": "Additional error data",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}